Update Github workflow build & run destinations, fix respective integration tests (#545)
All tests are updated to use macOS 15, Xcode 16.4, and iOS 18.6. This also requires an update to the integration tests, which were looking for a different title on the Password Manager field.

Ordinarily, I would see "breakage due to macos-latest/default sims/default xcode" changing as a useful reminder to update our repo. Due to [an unfixed bug in the iOS 18.4 simulators](https://developer.apple.com/forums/thread/777999?page=4) and https://github.com/actions/runner-images/issues/12758, using macOS-latest and the default simulators does not work.

More on Github runtime updates:
Github has https://github.com/actions/runner-images/issues/12541 the latest three versions of the platform tools for Xcode. If the default Xcode runtime for a macOS image is not one of those, simulators won't be available. This is currently true of macOS-15, where Xcode 16.0 is the default. As such, we need to specify the Xcode version. I've specified the macOS version as well to ensure that version of Xcode is available.